# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: takos API
  version: "1.0.0"
  description: |
    takos のバックエンド API を OpenAPI 形式で記述した仕様書です。
    認証はセッション Cookie を利用します。
servers:
  - url: https://{domain}
    description: 本番環境
    variables:
      domain:
        default: example.com
  - url: http://localhost:8000
    description: ローカル開発環境
paths:
...
  /api/posts:
    get:
      summary: 投稿一覧取得
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: before
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 投稿の配列
    post:
      summary: 新規投稿
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                author:
                  type: string
                content:
                  type: string
              required:
                - author
                - content
      responses:
        "201":
          description: 作成された投稿
  /api/posts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: 投稿取得
      responses:
        "200":
          description: 投稿
    put:
      summary: 投稿更新
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
              required:
                - content
      responses:
        "200":
          description: 更新後の投稿
    delete:
      summary: 投稿削除
      responses:
        "200":
          description: 成功
...
  /api/posts/{id}/like:
    post:
      summary: いいね
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
              required:
                - username
      responses:
        "200":
          description: いいね数
  /api/posts/{id}/retweet:
    post:
      summary: リツイート
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
              required:
                - username
      responses:
        "200":
          description: リツイート数
...